requirements.txt
nose
sphinx
six
enum34
argparse
requests
Pillow
jsonpickle
pyyaml

numpy
scipy
scikit-image
scikit-learn
matplotlib

mock
tempdir